+++ internal/client/models.go
@@ -116,6 +116,8 @@ type WebDomain struct {
 	PMMaxRequests     FlexInt `json:"pm_max_requests,omitempty"`
 	HTTPPort          FlexInt `json:"http_port,omitempty"`
 	HTTPSPort         FlexInt `json:"https_port,omitempty"`
+	PHPOpenBasedir    string  `json:"php_open_basedir,omitempty"`
+	ApacheDirectives  string  `json:"apache_directives,omitempty"`
 	CustomPHPIni      string  `json:"custom_php_ini,omitempty"`
 	BackupInterval    string  `json:"backup_interval,omitempty"`
 	BackupCopies      FlexInt `json:"backup_copies,omitempty"`